extracted assets from Reader to separate class to reduce Reader class definition
